xfs/189: fix remount barrier test failure
_test_remount_barrier() is skipped on kernels newer than v4.19 where the
barrier options have been removed, and this is detected by matching the
mount command output for a "bad option" error.

Recent util-linux commit 9da5644e414c ("libmount: report kernel message
from new API"), which was introduced in v2.41 release changed that error
message from the old hardcoded one:

  mount: /mnt/scratch: mount point not mounted or bad option.

to the newer, and more specific coming from the kernel:

  mount: /mnt/scratch: fsconfig() failed: xfs: Unknown parameter 'barrier'.

Update the test to account for that change so that the unsupported
options can be properly detected again and for this part of the test to
be skipped successfully.

Signed-off-by: Anthony Iliopoulos <ailiop@suse.com>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>