ext4/060: Skip for dax devices on non-4k page sizes
This test tries to trigger an ext4 corruption by resizing past the
meta_bg size (8GB on 4k blocksize) in ext4. Since the test is sensistive
to the size and FS layout it hardcodes the 4kb blocksize ignoring user's
$MKFS_OPTIONS. While this is okay for most cases it fails for dax based
filesystems where system pagesize is non-4k.

One way to work past this is to make the test blocksize agnostic, but
since we still need the disk to be as big as the meta_bg size, this
means for blocksize=64kb filesystems we need a scratch disk of ~4TB
which is not feasible.

Hence, just skip the test if fsdax is used in a non-4k page size system.

Reported-by: Disha Goel <disgoel@linux.ibm.com>
Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>
Reviewed-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: Zorro Lang <zlang@kernel.org>